- Optimized the handling of custom success messages by adding a no_date_found parameter and customizing the success message based on its value. - Added more specific error messages and instructions for different types of errors. - Improved the readability of the code.

+++ main.py
@@ -8,6 +8,7 @@ import filedate
 from exif import Image
 
 count = Counter()
+overwrite = True
 
 SUPPORTED_FORMATS = [".jpg", ".jpeg", ".png"]
 current_folder = os.getcwd()
@@ -39,9 +40,12 @@ def print_final_count(counter):
     print(
         "\033[1;32m" + f'SUCCESS: {success_count} image{"s" if success_count != 1 else ""} have been processed and '
                        f'saved.' + "\033[0m")
-    print(
-        "\033[1;31m" + f'ERROR: {error_count} image{"s" if error_count != 1 else ""} already existed and were not '
-                       f'overwritten.' + "\033[0m")
+    print_error_message(
+        f'ERROR: {error_count} image{"s" if error_count != 1 else ""} already existed and were not overwritten.')
+
+
+def print_error_message(msg):
+    print(f"\033[1;31m{str(msg)}\033[0m")
 
 
 def extract_exif_date(file):
@@ -54,18 +58,31 @@ def extract_exif_date(file):
             return img.has_exif
 
 
-def copy_image(source, destination_name, custom_success_message=None):
+def copy_image(source, destination, no_date_found=False):
     try:
-        destination = os.path.join(destination_folder, destination_name)
+        destination = os.path.join(destination_folder, destination)
         os.makedirs(os.path.dirname(destination), exist_ok=True)
+        if os.path.exists(destination) and not overwrite:
+            raise FileExistsError(f"The file '{os.path.basename(destination)}' "
+                                  f"already exists at '{os.path.relpath(destination)}'.\n "
+                                  f"To overwrite the file, please enable the 'overwrite' option in the user interface.")
         shutil.copy2(source, destination)
-        if custom_success_message:
-            print(custom_success_message)
-        else:
-            print(f"{f'SUCCESS: {destination} image have been processed and saved.'}")
+        print(f"The file '{os.path.basename(source)}' has been copied to the directory '{os.path.dirname(destination)}'"
+              f"{' BUT no date was found in the image name.' if no_date_found else '.'}")
+
         return True
+    except FileExistsError as e:
+        print_error_message(e)
+        return False
     except Exception as e:
-        print("\033[1;31m" + f"An error occurred: {e} " + "\033[0m")
+        if "PermissionError" in str(e):
+            print_error_message(f"An error occurred with '{source}': {e}.\nPlease check if you have the permission "
+                                f"to access the destination folder.")
+        elif "No such file or directory" in str(e):
+            print_error_message(f"An error occurred with '{source}': {e}.\nPlease check if the source file path is "
+                                f"correct.")
+        else:
+            print_error_message(f"An error occurred with '{source}': {e}")
         return False
 
 
@@ -149,8 +166,8 @@ def exif_date_change(src_folder, dst_folder):
 
             destination = os.path.join(png_folder, file)
             if not date:
-                message = f"{f'SUCCESS: {destination} image have been '}" \
-                            f"processed and saved BUT no date was found in the image name."
+                message = f"{file} successfully copied to {os.path.relpath(destination)} BUT no date was " \
+                          f"found in the image name."
                 update_counter(copy_image(file, destination, message))
 
             elif date:
